// src/types/Rule.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuleId } from "./RuleId";
import type { RuleSpec } from "./RuleSpec";

/**
 * A backup rule.
 */
export type Rule = { 
/**
 * Row id.
 */
id: RuleId, 
/**
 * The rule's configuration.
 */
spec: RuleSpec, 
/**
 * When the rule was created, as a Unix timestamp.
 */
created_at: number, };